+++ packages/next/src/client/components/router-reducer/router-reducer-types.ts
@@ -2,7 +2,7 @@ import type { CacheNode, ScrollRef } from '../../../shared/lib/app-router-types'
 import type { FlightRouterState } from '../../../shared/lib/app-router-types'
 import type { NavigationSeed } from '../segment-cache/decode-server-response'
 import type { FetchServerResponseResult } from './fetch-server-response'
-import type { FreshnessPolicy } from './ppr-navigations'
+import type { FreshnessPolicy } from '../render-tree'
 
 export const ACTION_REFRESH = 'refresh'
 export const ACTION_NAVIGATE = 'navigate'
